inference/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.7.0
torch==2.3.0
transformers==4.50.0
peft==0.14.0
accelerate==1.2.1
google-cloud-storage==2.16.0
google-cloud-bigquery==3.25.0
sacrebleu==2.4.0
sentencepiece==0.2.0
huggingface_hub>=0.30.0